requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.13.3
bs4==0.0.2
cached-property==2.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
commonmark==0.9.1
debugpy==1.8.13
decorator==5.2.1
dictdiffer==0.9.0
distro==1.9.0
executing==2.2.0
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.1.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
notion==0.0.28
openai==1.66.3
packaging==24.2
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pyzmq==26.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.1
text-unidecode==1.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzlocal==5.3
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
Werkzeug==2.2.3
